Revert "fix: autoconnect and allow disconnect (#298)"
This reverts commit 8517f4de334a8d6c03db978e8e5e0bd3a5b98fbb.
